refactor `validity_proof` module into separate modules
