+++ zk-token-sdk/src/zk_token_elgamal/pod/sigma_proofs.rs
@@ -2,15 +2,12 @@
 
 #[cfg(not(target_os = "solana"))]
 use crate::sigma_proofs::{
+    batched_grouped_ciphertext_validity_proof::BatchedGroupedCiphertext2HandlesValidityProof as DecodedBatchedGroupedCiphertext2HandlesValidityProof,
     ciphertext_ciphertext_equality_proof::CiphertextCiphertextEqualityProof as DecodedCiphertextCiphertextEqualityProof,
     ciphertext_commitment_equality_proof::CiphertextCommitmentEqualityProof as DecodedCiphertextCommitmentEqualityProof,
-    errors::*,
-    fee_proof::FeeSigmaProof as DecodedFeeSigmaProof,
+    errors::*, fee_proof::FeeSigmaProof as DecodedFeeSigmaProof,
+    grouped_ciphertext_validity_proof::GroupedCiphertext2HandlesValidityProof as DecodedGroupedCiphertext2HandlesValidityProof,
     pubkey_proof::PubkeyValidityProof as DecodedPubkeyValidityProof,
-    validity_proof::{
-        BatchedGroupedCiphertext2HandlesValidityProof as DecodedBatchedGroupedCiphertext2HandlesValidityProof,
-        GroupedCiphertext2HandlesValidityProof as DecodedGroupedCiphertext2HandlesValidityProof,
-    },
     zero_balance_proof::ZeroBalanceProof as DecodedZeroBalanceProof,
 };
 use crate::zk_token_elgamal::pod::{Pod, Zeroable};
